Add files via upload
The program mimic a card game called with 52 cards in the deck. The player and the computer are each dealt five cards from the standard French deck. In this deck, Ace is worth 11 points, J, Q, K are each worth 10 points and the others are worth their original value. The cards and total points of the player's hand are displayed. The player then decides whether she thinks the computer has a better hand or not. After the player's input, the computer's hand and points are displayed. The player wins or loses depending on whether she guessed the outcome correctly. Play continues until the player quits or there are no more cards in the deck.